add test__dvf_modeule, migrated from previous file structure in saf_test.c
