+++ test/include/saf_test.h
@@ -366,6 +366,21 @@ void test__saf_example_spreader(void);
 #endif /* SAF_ENABLE_EXAMPLES_TESTS */
 
 
+/* ========================================================================== */
+/*                           SAF DVF module unit tests                          */
+/* ========================================================================== */
+
+/**
+ * Calculate high shelf parameters, g0, gInf, fc, from the lookup table coefficients (10 degree steps). */
+void test__dvf_calcHighShelfParams(void);
+/**
+ * Test the interpolation of high shelf parameters based on distance and incidence angle parameters */
+void test__dvf_interpHighShelfParams(void);
+/**
+ * Test thegeneration of high shelf coeffs based on shelf gains and fc parameters */
+void test__dvf_calcIIRCoeffs(void);
+
+
 #ifdef __cplusplus
 } /* extern "C" */
 #endif /* __cplusplus */
